refactor: extract location filter logic to a view model
- Moves state management and business logic from LocationFilterView to a new LocationFilterViewModel to improve code structure and testability.
- Creates LocationFilterViewModel to handle the logic for location filtering.
- Relocates the LocationPreset enum from the view to the new view model, as it is part of the feature's business logic.
- Replaces local @State variables in the view with a single @StateObject reference to the view model.
- Simplifies the view by delegating all user actions, such as applying presets and clearing the filter, to the view model.
